Create empty `bip158` crate
As part of discussion #5331, `bitcoin` and `p2p` should not depend on
each other and should instead mutually depend on `primitives`, `hashes`,
etc. BIP-158 is the last remaining dependency from `bitcoin` in `p2p`.
It is not a peer-to-peer specification, but it is not an essential
module of `bitcoin` either. IMO the best option for this module is to
release it as a crate.

Blaming the module shows it hasn't changed architecturally in many
years. A new crate would allow us to revisit some of the design choices.
Particularly the heavy use of generics, dependency on `io` that could be
replaced by `consensus_encoding`, and perhaps an opportunity to benchmark
and improve performance.